Conflicts:
	EgammaAnalysis/ElectronTools/interface/ElectronEnergyCalibratorRun2.h
	EgammaAnalysis/ElectronTools/interface/PhotonEnergyCalibratorRun2.h
	EgammaAnalysis/ElectronTools/interface/SimpleElectron.h
	EgammaAnalysis/ElectronTools/plugins/CalibratedElectronProducersRun2.cc
	EgammaAnalysis/ElectronTools/plugins/CalibratedPhotonProducersRun2.cc
	EgammaAnalysis/ElectronTools/python/calibratedElectronsRun2_cfi.py
	EgammaAnalysis/ElectronTools/python/calibratedPhotonsRun2_cfi.py
	EgammaAnalysis/ElectronTools/src/ElectronEnergyCalibratorRun2.cc
	EgammaAnalysis/ElectronTools/src/PhotonEnergyCalibratorRun2.cc
	EgammaAnalysis/ElectronTools/src/SimpleElectron.cc
	EgammaAnalysis/ElectronTools/src/SimplePhoton.cc
	EgammaAnalysis/ElectronTools/src/classes.h
	EgammaAnalysis/ElectronTools/src/classes_def.xml
	EgammaAnalysis/ElectronTools/test/testCalibratedPatElectronProducerRun2_cfg.py

+++ EgammaAnalysis/ElectronTools/interface/SimpleElectron.h
@@ -9,7 +9,7 @@ class SimpleElectron
 {
 	public:
 	    SimpleElectron(){}
-	    SimpleElectron( unsigned int run, 
+	    SimpleElectron( double run, 
                         double eClass, 
                         double r9, 
                         double scEnergy, 
@@ -69,7 +69,7 @@ class SimpleElectron
     	double getEta() const {return eta_;}
     	float getR9() const {return r9_;}
     	int getElClass() const {return eClass_;}
-    	unsigned int getRunNumber() const {return run_;}
+    	int getRunNumber() const {return run_;}
     	bool isEB() const {return isEB_;}
     	bool isMC() const {return isMC_;}
     	bool isEcalDriven() const {return isEcalDriven_;}
@@ -82,7 +82,7 @@ class SimpleElectron
     	void setNewEnergyError(double newEnergyError){newEnergyError_ = newEnergyError;}
 
 	private:
-    	unsigned int run_; 
+    	double run_; 
         double eClass_;
     	double r9_;
     	double scEnergy_;
